chore(arena): move conversation.py to ask.py
